Some fixes and cleanup Some tests were using wrong for of DeqIO and EnqIO moved options stuff into it's own file and created a wrapper class for options management makes everything look cleaner add example to gcd on how to call
