+++ src/main/scala/chisel3/iotesters/Driver.scala
@@ -5,71 +5,10 @@ package chisel3.iotesters
 import java.io.File
 
 import chisel3.{HasChiselExecutionOptions, Module}
-import firrtl.{ComposableOptions, ExecutionOptionsManager, HasFirrtlOptions}
+import firrtl.{ExecutionOptionsManager, HasFirrtlOptions}
 import firrtl_interpreter.HasInterpreterOptions
 
-import scala.collection.mutable
 import scala.util.DynamicVariable
-
-
-case class TesterOptions(
-   isGenVerilog:    Boolean = false,
-   isGenHarness:    Boolean = false,
-   isCompiling:     Boolean = false,
-   isRunTest:       Boolean = false,
-   testerSeed:      Long    = System.currentTimeMillis,
-   testCmd:         mutable.ArrayBuffer[String]= mutable.ArrayBuffer[String](),
-   backendName:     String  = "firrtl",
-   logFileName:     String  = "",
-   waveform:        Option[File] = None) extends ComposableOptions
-
-trait HasTesterOptions {
-  self: ExecutionOptionsManager =>
-
-  var testerOptions = TesterOptions()
-
-  parser.note("tester options")
-
-  parser.opt[String]("backend-name")
-    .abbr("tbn")
-    .foreach { x => testerOptions = testerOptions.copy(backendName = x) }
-    .text("run this as test command")
-
-  parser.opt[Unit]("is-gen-verilog")
-    .abbr("tigv")
-    .foreach { _ => testerOptions = testerOptions.copy(isGenVerilog = true) }
-    .text("has verilog already been generated")
-
-  parser.opt[Unit]("is-gen-harness")
-    .abbr("tigh")
-    .foreach { _ => testerOptions = testerOptions.copy(isGenHarness = true) }
-    .text("has harness already been generated")
-
-  parser.opt[Unit]("is-compiling")
-    .abbr("tic")
-    .foreach { _ => testerOptions = testerOptions.copy(isCompiling = true) }
-    .text("has harness already been generated")
-
-  parser.opt[Seq[String]]("test-command")
-    .abbr("ttc")
-    .foreach { x => testerOptions = testerOptions.copy(testCmd = testerOptions.testCmd ++ x) }
-    .text("run this as test command")
-
-  parser.opt[String]("log-file-name")
-    .abbr("tlfn")
-    .foreach { x => testerOptions = testerOptions.copy(logFileName = x) }
-    .text("write log file")
-
-  parser.opt[File]("wave-form-file-name")
-    .abbr("twffn")
-    .foreach { x => testerOptions = testerOptions.copy(waveform = Some(x)) }
-    .text("wave form file name")
-
-  parser.opt[Long]("test-seed")
-    .abbr("tts")
-    .foreach { x => testerOptions = testerOptions.copy(testerSeed = x) }
-    .text("provides a seed for random number generator")
-}
 
 object Driver {
   private val backendVar = new DynamicVariable[Option[Backend]](None)
@@ -77,11 +16,7 @@ object Driver {
 
   def execute[T <: Module](
                             dutGenerator: () => T,
-                            optionsManager: ExecutionOptionsManager
-                              with HasTesterOptions
-                              with HasChiselExecutionOptions
-                              with HasFirrtlOptions
-                              with HasInterpreterOptions
+                            optionsManager: TesterOptionsManager
                           )
                           (
                             testerGen: T => PeekPokeTester[T]
@@ -117,11 +52,7 @@ object Driver {
   def execute[T <: Module](args: Array[String], dut: () => T)(
     testerGen: T => PeekPokeTester[T]
   ): Boolean = {
-    val optionsManager = new ExecutionOptionsManager("chisel-testers")
-      with HasTesterOptions
-      with HasChiselExecutionOptions
-      with HasFirrtlOptions
-      with HasInterpreterOptions
+    val optionsManager = new TesterOptionsManager
 
     optionsManager.parse(args) match {
       case true =>
@@ -148,11 +79,7 @@ object Driver {
     */
   def apply[T <: Module](dutGen: () => T, backendType: String = "firrtl")(
       testerGen: T => PeekPokeTester[T]): Boolean = {
-    val optionsManager = new ExecutionOptionsManager("chisel-testers")
-      with HasTesterOptions
-      with HasChiselExecutionOptions
-      with HasFirrtlOptions
-      with HasInterpreterOptions
+    val optionsManager = new TesterOptionsManager
 
     val (dut, backend) = backendType match {
       case "firrtl" => setupFirrtlTerpBackend(dutGen, optionsManager)
