+++ src/test/scala/examples/SmallOdds4.scala
@@ -9,8 +9,8 @@ import chisel3.iotesters.{ChiselFlatSpec, OrderedDecoupledHWIOTester}
 class SmallOdds4(filter_width: Int) extends Module {
 
   class FilterIO extends Bundle {
-    val in = new DeqIO(UInt(width = filter_width))
-    val out = new EnqIO(UInt(width = filter_width))
+    val in = DeqIO(UInt(width = filter_width))
+    val out = EnqIO(UInt(width = filter_width))
   }
 
   class Filter(isOk: UInt => Bool) extends Module {
